% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_2_calc_distances.R
\name{calc_circle_distance}
\alias{calc_circle_distance}
\title{calc_circle_distance}
\usage{
calc_circle_distance(spatial_df)
}
\description{
a function that calculates the lead distances of vessel
observations. calculate a new tibble after applied function
}
