% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_create_map.R
\name{create_distance_map_between_a_n_b}
\alias{create_distance_map_between_a_n_b}
\title{create_distance_map_between_a_n_b}
\usage{
create_distance_map_between_a_n_b(filtered_df)
}
\description{
may function that calculates distances
and displays results in a map
}
