% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_filter_vessel_name.R
\name{filter_vessel_name}
\alias{filter_vessel_name}
\title{filter_vessel_name}
\usage{
filter_vessel_name(df, vessel_name)
}
\description{
function to filter vessel records by name
import dplyr
}
