% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_filter_vesselname_by_type.R
\name{filter_vessel_type}
\alias{filter_vessel_type}
\title{filter_vessel_type}
\usage{
filter_vessel_type(df, vessel_type)
}
\description{
function to filter vessel names n typs
by type. used for dynamic filtering in dropdown module
}
