% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_max_distance.R
\name{get_max_distance}
\alias{get_max_distance}
\title{get_max_distance}
\usage{
get_max_distance(df_distances)
}
\description{
function to obtain a single record regarding
the most early longest consecutive distance
}
