% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-leafletmap.R
\name{leaftlet_map_server}
\alias{leaftlet_map_server}
\title{ship_dropdowns_server}
\usage{
leaftlet_map_server(id, dropdown_data)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}

\item{filtered_df}{data frame (reactive) with variables necessary for leaflet map}
}
\value{
list with following components
\describe{
\item{leaflet_map}{reactive leaflet map to be displayed}
}
}
\description{
ship_dropdowns_server
}
