% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_load_ship_names_n_types.R
\name{load_ship_names_n_types}
\alias{load_ship_names_n_types}
\title{load_ship_names_n_types}
\usage{
load_ship_names_n_types()
}
\description{
load csv data for data selector dynamic filtering
}
