% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_record_of_top_distance.R
\name{obtain_record_of_top_distance}
\alias{obtain_record_of_top_distance}
\title{obtain_record_of_top_distance}
\usage{
obtain_record_of_top_distance(filtered_df, top_distance)
}
\description{
function to obtain a df with observations of 2
records initial point and movement to point
}
