% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dropdowns.R
\name{ship_dropdowns_server}
\alias{ship_dropdowns_server}
\title{ship_dropdowns_server}
\usage{
ship_dropdowns_server(id)
}
\arguments{
\item{input, }{output, session standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
\item{my_filtered_df}{reactive tibble used for performing calculation the create_distance_map_between_a_n_b function}
\item{vessel_name}{reactive character of vessel selected}
}
}
\description{
ship_dropdowns_server
}
