% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dropdowns.R
\docType{data}
\name{ship_names_n_types}
\alias{ship_names_n_types}
\title{Dropdown module 4 filtering vessels}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1226 rows and 2 columns.
}
\usage{
ship_names_n_types
}
\description{
the module has 2 selectors and loads the needed
data for the filtering of the big data frame
}
\keyword{datasets}
